move view models to separate namespace
