+++ src/CSharp App/View Structures/SettingsViewPersistentCache.cs
