+++ src/CSharp App/ViewModels/DialogWindowViewModel.cs
