+++ src/CSharp App/ViewModels/MainWindowViewModel.cs
