+++ src/CSharp App/ViewModels/SettingsViewModel.cs
@@ -5,7 +5,6 @@ using Avalonia.Media.Imaging;
 using VolumetricSelection2077.Resources;
 using VolumetricSelection2077.Services;
 using VolumetricSelection2077.Views;
-using YamlDotNet.Core.Tokens;
 
 namespace VolumetricSelection2077.ViewModels
 {
