+++ src/CSharp App/Views/DebugWindow.axaml
@@ -8,12 +8,13 @@
     d:DesignWidth="800"
     mc:Ignorable="d"
     x:Class="VolumetricSelection2077.Views.DebugWindow"
-    x:DataType="viewmodels:DebugWindowViewModel"
+    x:DataType="viewModels:DebugWindowViewModel"
     xmlns="https://github.com/avaloniaui"
     xmlns:avalonia="clr-namespace:LoadingIndicators.Avalonia;assembly=LoadingIndicators.Avalonia"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:resources="clr-namespace:VolumetricSelection2077.Resources"
+    xmlns:viewModels="clr-namespace:VolumetricSelection2077.ViewModels"
     xmlns:viewmodels="clr-namespace:VolumetricSelection2077.Views"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <Grid
